added priority part 6 (design task_details_screen)
