+++ lib/features/task_details/widgets/in_progress_indicator_widget.dart
