+++ lib/features/task_details/widgets/widgets.dart
@@ -1,2 +1,4 @@
-export 'in_progress_indicator.dart';
+export 'in_progress_indicator_widget.dart';
 export 'weather_widget.dart';
+export 'main_information_widget.dart';
+export 'overview_widget.dart';
